フロントエンドの構成を整理
