+++ frontend/react-app/src/util.js
@@ -1,4 +1,5 @@
 import { assoc } from 'ramda';
+import * as vis from 'vis';
 
 export const toSnakeCase = (str) => {
   return str.split(/(?=[A-Z])/).join('_').toLowerCase();
@@ -24,4 +25,41 @@ export const createRequestBody = (arr) => {
 
   // スネークケースに変換
   return toSnakeCaseObject(obj)
+}
+
+export const drawGraph = (nodes, edges) => {
+  const container = document.getElementById('network');
+  const data = {
+    nodes: nodes,
+    edges: edges
+  };
+  const options = {
+    edges: {
+      chosen: false
+    },
+    // todo: ここをうまいこといじると、リンクの重複を解消できそう
+    // 要ドキュメントよみこみ https://visjs.github.io/vis-network/docs/network/
+    // physics: {
+    //   enabled: true,
+    //   hierarchicalRepulsion: {
+    //     centralGravity: 0.0,
+    //     springLength: 500,
+    //     springConstant: 0.01,
+    //     nodeDistance: 100,
+    //     damping: 0.09
+    //   },
+    //   solver: 'hierarchicalRepulsion'
+    // },
+    interaction: {
+      dragView: false,
+      zoomView: false
+    },
+    layout: {
+      randomSeed: 0
+    },
+    nodes: {
+      chosen: false
+    },
+  };
+  new vis.Network(container, data, options);
 }
